ai_work_proxy/requirements.txt
google-generativeai
sqlite3